Crate node_cli

Substrate CLI library.

This package has two Cargo features:

cli (default): exposes functions that parse command-line options, then start and run the node as a CLI application.

browser: exposes the content of the browser module, which consists of exported symbols that are meant to be passed through the wasm-bindgen utility and called from JavaScript. Despite its name the produced WASM can theoretically also be used from NodeJS, although this hasn’t been tested.

Modules

chain_spec
Substrate chain configurations.

Structs

Cli
An overarching CLI command definition.

Enums

Subcommand
Possible subcommands of the main binary.

Functions

__wasm_bindgen_generated_start_client
Starts the client.

run
Parse command line arguments into service configuration.

start_client
Starts the client.