Struct node_cli::chain_spec::Extensions

pub struct Extensions {
    pub fork_blocks: ForkBlocks<Block>,
    pub bad_blocks: BadBlocks<Block>,
    pub light_sync_state: LightSyncStateExtension,
}

Node ChainSpec extensions.

Additional parameters for some Substrate core modules, customizable from the chain spec.

Fields
fork_blocks: ForkBlocks<Block>

Block numbers with known hashes.

bad_blocks: BadBlocks<Block>

Known bad block hashes.

light_sync_state: LightSyncStateExtension

The light sync state extension used by the sync-state rpc.